Run jupyter ipynb notebooks via browser-plus on 0.0.0.0:8888
Command to run should now be docker run --rm -it -v ... ahb. Have a start_script.sh to handle jupyter notebook server start and atom-beta launch. The jupyter token should be visible in the terminal on running the `docker run ...` command.

Put placeholder for 'Tini' in case that might be needed in the future as prefix to the jupyter notebook init sh script.