:construction: Experimental work around holoviews linked streams
Wanting to get a working interactive notebook, where a user can zoom in on a bounding box area of Antarctica, and see the corresponding ICESAT elevation points of the lasers plotted in real time.

Loosely based on https://anaconda.org/petrenko/linking_datashaders/notebook and https://stackoverflow.com/questions/44287203/how-to-do-linked-data-selections-in-holoviews-with-datashader-bokeh-backend plus a few other scattered holoview tutorials over the web.

WIll be putting this on hold for a while as fighting with the plotting APIs has not been productive. Moving on to other side work e.g. vectorized point in polygon stuff.