:fire: Nearly full shift from pip to conda, revert geopandas-cython
Purpose of this commit is to move from mixing pip and conda to a more 'true' conda environment. Kinda helps when you want to use a high performance computing linux server environment that doesn't let you install docker and or use sudo so much. Transition is made evident with the replacement of the (long empty...) requirements.txt with an environment.yml file (that needs to be less hardcoded). Much more work needed to restructure this project!!

Note that we are using conda version 4.4.0 to make use of newer syntax for installing packages from particular conda channels. I.e. we can do `conda install bokeh::datashader`. This is particularly useful for the datashader, geoviews and holoviews packages.

Also temporarily reverting the use of geopandas-cython branch. There are a few known repercussions in the h5_to_np_icesat.ipynb I'll need to fix but this is low priority now.