:building_construction: Pivoting to Deep Learning in the Cryosphere
Change of title from a long windy detailed one to a short and to the point one. Added a more scientific rationale on why we're doing this project, and made the previous 'background' into a 'plan' instead.

Gave a short introduction to artificial neural networks which is a replacement of the in depth convnet yarn I had before the great purge. Put in a 3 by 4 grid of images of some of the datasets we'll be using courtesy of Quantarctica3.

Added some more detail to each of the 3 sections in the project outline. Declared funding sources and resource requirements, and acknowledged some of the courses and competitions I have taken part in for my personal learning. Purged the summary (should have done this in the big purge...).