:memo: Work on adding basal slip to background, and misc cleanup
Mostly adding more words to the introductory background section on basal slip which is basically basal sliding + bed deformation. In other parts, continued to cleanup some unnecessary cruft such as the precise geographical coverage of this project (i.e. just Antarctica really) and the technicalities of the neural network model used (who cares what type of ConvNet it is anyway?).

There was some work on adding SVG figures of neural network models in the Topic Development/Background and Methodology/neural network model section but had some issues with Pandoc rendering it so leaving it out for now.